feat: split codebase on subclasses
